Solve day 15
I'm a bit dissatisfied with the brute force solution to part 2, which
for hours I was reluctant to try.  Perhaps I shouldn't have been
so reluctant, since it only ran for 11 seconds when I finally tried
it, and in less than 1 second after compiling in release mode 😅

Still, I feel like there ought to be a much smarter and more elegant
solution.